Move FS based func to internal/fs module.
