+++ merge/merge.go
@@ -5,11 +5,10 @@ import (
 	"io"
 	"os"
 	"path/filepath"
-	"sort"
 	"time"
 
 	internaldb "github.com/wtsi-ssg/wrstat/v4/internal/db"
-	ifs "github.com/wtsi-ssg/wrstat/v4/internal/fs"
+	"github.com/wtsi-ssg/wrstat/v4/internal/fs"
 )
 
 const (
@@ -18,7 +17,7 @@ const (
 )
 
 func Merge(sourceDir, destDir string, removeOld bool) error {
-	de, err := FindLatestCombinedOutputOlderThan(sourceDir, reloadGrace)
+	de, err := fs.FindLatestCombinedOutputOlderThan(sourceDir, watchFile, reloadGrace)
 	if err != nil {
 		return fmt.Errorf("failed to find database files in source dir: %w", err)
 	}
@@ -41,60 +40,6 @@ func Merge(sourceDir, destDir string, removeOld bool) error {
 	}
 
 	return nil
-}
-
-type pathTime struct {
-	path    string
-	modtime time.Time
-}
-
-// FindLatestCombinedOutputOlderThan finds the latest entry in dir and returns its path.
-func FindLatestCombinedOutputOlderThan(dir string, minAge time.Duration) (string, error) {
-	for {
-		files, err := filepath.Glob(filepath.Join(dir, "*", "*", "*", watchFile))
-		if err != nil {
-			return "", err
-		}
-
-		if len(files) == 0 {
-			return "", ifs.ErrNoDirEntryFound
-		}
-
-		de, err := filesToLatestPathTime(files)
-		if err != nil {
-			return "", err
-		}
-
-		diff := de.modtime.Sub(time.Now().Add(-minAge))
-
-		if diff < 0 {
-			return filepath.Dir(filepath.Dir(filepath.Dir(de.path))), nil
-		}
-
-		time.Sleep(diff)
-	}
-}
-
-func filesToLatestPathTime(files []string) (pathTime, error) {
-	des := make([]pathTime, len(files))
-
-	for n, file := range files {
-		fi, err := os.Lstat(file)
-		if err != nil {
-			return pathTime{}, err
-		}
-
-		des[n] = pathTime{
-			path:    file,
-			modtime: fi.ModTime(),
-		}
-	}
-
-	sort.Slice(des, func(i, j int) bool {
-		return des[i].modtime.After(des[j].modtime)
-	})
-
-	return des[0], nil
 }
 
 func copyPreservingTimestamp(source, dest string) error {
